refactor(web): split recordings components
