+++ apps/web/src/components/recordings/recording-analysis-page.tsx
@@ -4,22 +4,14 @@ import { toast } from 'sonner';
 import { useSuspenseQuery } from '@tanstack/react-query';
 import { useNavigate } from '@tanstack/react-router';
 
-import type { Recording, RecordingAnalysis } from '@stitch/shared/recordings/types';
-
 import { AnalysisHeader } from './analysis/analysis-header';
+import { buildAnalysisMarkdown } from './analysis/build-analysis-markdown';
 import { SummarySection } from './analysis/summary-section';
 import { TopicList } from './analysis/topic-list';
 import { TranscriptSidebar } from './analysis/transcript-sidebar';
+import { getErrorMessage, shouldConfirmRecordingDelete } from './shared/actions';
+import { DeleteRecordingDialog } from './shared/delete-recording-dialog';
 
-import { Button } from '@/components/ui/button';
-import {
-  Dialog,
-  DialogContent,
-  DialogDescription,
-  DialogFooter,
-  DialogHeader,
-  DialogTitle,
-} from '@/components/ui/dialog';
 import { ScrollArea } from '@/components/ui/scroll-area';
 import {
   recordingAnalysisQueryOptions,
@@ -29,109 +21,6 @@ import {
   useStartRecordingAnalysis,
   useStopRecording,
 } from '@/lib/queries/recordings';
-
-const ACTION_STATUS_LABEL: Record<string, string> = {
-  todo: 'To do',
-  in_progress: 'In progress',
-  done: 'Done',
-  unknown: 'Unknown',
-};
-
-function pushList(lines: string[], title: string, items: string[]): void {
-  if (!items.length) return;
-  lines.push(`**${title}**`);
-  lines.push(...items.map((item) => `- ${item}`));
-  lines.push('');
-}
-
-function buildAnalysisMarkdown(
-  analysis: RecordingAnalysis | null | undefined,
-  recording: Recording | undefined,
-): string | null {
-  if (!analysis || analysis.status !== 'completed') return null;
-
-  const lines: string[] = [];
-  const title = analysis.title || recording?.title || 'Recording analysis';
-
-  lines.push(`# ${title}`);
-  lines.push('');
-
-  if (analysis.summary.trim()) {
-    lines.push('## Executive Summary');
-    lines.push('');
-    lines.push(analysis.summary.trim());
-    lines.push('');
-  }
-
-  if (analysis.topicSections.length) {
-    lines.push('## Topic Analysis');
-    lines.push('');
-
-    analysis.topicSections.forEach((section, index) => {
-      lines.push(`### ${index + 1}. ${section.name}`);
-      lines.push('');
-      lines.push(`**Turns:** ${section.startTurn + 1}-${section.endTurn + 1}`);
-      lines.push('');
-
-      if (section.analysis.trim()) {
-        lines.push(section.analysis.trim());
-        lines.push('');
-      }
-
-      pushList(lines, 'Decisions', section.decisions);
-
-      if (section.actionItems.length) {
-        const actionItems = section.actionItems.map((item) => {
-          const metadata = [
-            `assignee: ${item.assignee ?? 'Unassigned'}`,
-            item.dueDate ? `due: ${item.dueDate}` : null,
-            `status: ${ACTION_STATUS_LABEL[item.status] ?? item.status}`,
-          ].filter(Boolean);
-
-          return `${item.task} (${metadata.join('; ')})`;
-        });
-
-        pushList(lines, 'Action Items', actionItems);
-      }
-
-      if (section.blockers.length) {
-        const blockers = section.blockers.map((blocker) => {
-          const metadata = [
-            `assignee: ${blocker.assignee ?? 'Unassigned'}`,
-            `impact: ${blocker.impact ?? 'Unknown'}`,
-          ];
-          return `${blocker.description} (${metadata.join('; ')})`;
-        });
-
-        pushList(lines, 'Risks & Blockers', blockers);
-      }
-
-      pushList(lines, 'Open Questions', section.openQuestions);
-      pushList(lines, 'Next Steps', section.nextSteps);
-    });
-  }
-
-  if (analysis.blockers.length) {
-    const blockers = analysis.blockers.map((blocker) => {
-      const metadata = [
-        `assignee: ${blocker.assignee ?? 'Unassigned'}`,
-        `impact: ${blocker.impact ?? 'Unknown'}`,
-      ];
-
-      return `${blocker.description} (${metadata.join('; ')})`;
-    });
-
-    lines.push('## Overall Risks & Blockers');
-    lines.push('');
-    lines.push(...blockers.map((blocker) => `- ${blocker}`));
-    lines.push('');
-  }
-
-  return lines
-    .join('\n')
-    .replace(/\n{3,}/g, '\n\n')
-    .trim();
-}
 
 export function RecordingAnalysisPage({ recordingId }: { recordingId: string }) {
   const { data: analysisResponse } = useSuspenseQuery(recordingAnalysisQueryOptions(recordingId));
@@ -154,11 +43,21 @@ export function RecordingAnalysisPage({ recordingId }: { recordingId: string }) 
 
   const isRunning = analysis?.status === 'processing';
 
+  const deleteRecordingById = React.useCallback(() => {
+    void deleteRecording.mutateAsync(recordingId).then(
+      () => {
+        setShowDeleteConfirm(false);
+        toast.success('Recording deleted');
+        void navigate({ to: '/recordings' });
+      },
+      (error: unknown) => toast.error(getErrorMessage(error, 'Failed to delete recording')),
+    );
+  }, [deleteRecording, navigate, recordingId]);
+
   const handleStartAnalysis = () => {
     void startAnalysis.mutateAsync({ recordingId, force: true }).then(
       () => toast.success('Analysis started'),
-      (error: unknown) =>
-        toast.error(error instanceof Error ? error.message : 'Failed to start recording analysis'),
+      (error: unknown) => toast.error(getErrorMessage(error, 'Failed to start recording analysis')),
     );
   };
 
@@ -166,13 +65,12 @@ export function RecordingAnalysisPage({ recordingId }: { recordingId: string }) 
     void cancelAnalysis.mutateAsync(recordingId).then(
       () => toast.success('Analysis cancelled'),
       (error: unknown) =>
-        toast.error(error instanceof Error ? error.message : 'Failed to cancel recording analysis'),
+        toast.error(getErrorMessage(error, 'Failed to cancel recording analysis')),
     );
   };
 
   return (
     <div className="mx-auto flex h-full w-full max-w-400 flex-col gap-6 overflow-hidden px-4 py-6 sm:px-6 lg:px-8">
-      {/* Header */}
       <AnalysisHeader
         analysis={analysis}
         analysisMarkdown={analysisMarkdown}
@@ -188,27 +86,16 @@ export function RecordingAnalysisPage({ recordingId }: { recordingId: string }) 
         onStopRecording={() => {
           void stopRecording.mutateAsync().then(
             () => toast.success('Recording stopped'),
-            (error: unknown) =>
-              toast.error(error instanceof Error ? error.message : 'Failed to stop recording'),
+            (error: unknown) => toast.error(getErrorMessage(error, 'Failed to stop recording')),
           );
         }}
         onDelete={() => {
-          if (
-            recording?.durationMs !== null &&
-            recording?.durationMs !== undefined &&
-            recording.durationMs <= 30_000
-          ) {
-            void deleteRecording.mutateAsync(recordingId).then(
-              () => {
-                toast.success('Recording deleted');
-                void navigate({ to: '/recordings' });
-              },
-              (error: unknown) =>
-                toast.error(error instanceof Error ? error.message : 'Failed to delete recording'),
-            );
-          } else {
+          if (!recording || shouldConfirmRecordingDelete(recording)) {
             setShowDeleteConfirm(true);
+            return;
           }
+
+          deleteRecordingById();
         }}
       />
 
@@ -218,11 +105,9 @@ export function RecordingAnalysisPage({ recordingId }: { recordingId: string }) 
         </div>
       ) : null}
 
-      {/* Main Layout Grid */}
       <div className="grid min-h-0 flex-1 grid-cols-1 gap-6 lg:grid-cols-12">
-        {/* Left Column - Summary & Topics (Scrollable) */}
         <div className="flex min-h-0 flex-col lg:col-span-8 xl:col-span-8 2xl:col-span-9">
-          <ScrollArea className="flex-1 rounded-xl" style={{ height: 0 }}>
+          <ScrollArea className="h-0 flex-1 rounded-xl">
             <div className="space-y-8 pr-2 pb-12">
               <SummarySection analysis={analysis} isRunning={isRunning} />
               <TopicList sections={analysis?.topicSections} isRunning={isRunning} />
@@ -230,7 +115,6 @@ export function RecordingAnalysisPage({ recordingId }: { recordingId: string }) 
           </ScrollArea>
         </div>
 
-        {/* Right Column - Full Height Transcript */}
         <div className="min-h-0 lg:col-span-4 xl:col-span-4 2xl:col-span-3">
           <TranscriptSidebar
             analysis={analysis}
@@ -241,43 +125,12 @@ export function RecordingAnalysisPage({ recordingId }: { recordingId: string }) 
         </div>
       </div>
 
-      <Dialog
-        open={showDeleteConfirm}
+      <DeleteRecordingDialog
+        recording={showDeleteConfirm ? recording : null}
+        isDeleting={deleteRecording.isPending}
         onOpenChange={(open) => !open && setShowDeleteConfirm(false)}
-      >
-        <DialogContent>
-          <DialogHeader>
-            <DialogTitle>Delete recording?</DialogTitle>
-            <DialogDescription>
-              This permanently deletes &quot;{recording?.title}&quot; and its local audio file.
-            </DialogDescription>
-          </DialogHeader>
-          <DialogFooter>
-            <Button variant="outline" onClick={() => setShowDeleteConfirm(false)}>
-              Cancel
-            </Button>
-            <Button
-              variant="destructive"
-              onClick={() => {
-                void deleteRecording.mutateAsync(recordingId).then(
-                  () => {
-                    setShowDeleteConfirm(false);
-                    toast.success('Recording deleted');
-                    void navigate({ to: '/recordings' });
-                  },
-                  (error: unknown) =>
-                    toast.error(
-                      error instanceof Error ? error.message : 'Failed to delete recording',
-                    ),
-                );
-              }}
-              disabled={deleteRecording.isPending}
-            >
-              {deleteRecording.isPending ? 'Deleting...' : 'Delete'}
-            </Button>
-          </DialogFooter>
-        </DialogContent>
-      </Dialog>
+        onConfirm={deleteRecordingById}
+      />
     </div>
   );
 }
